+++ src/app/routes/performances.ts
@@ -5,13 +5,17 @@ import * as ttts from '@tokyotower/domain';
 import * as express from 'express';
 import { query } from 'express-validator';
 import { NO_CONTENT } from 'http-status';
-import * as moment from 'moment';
+import * as moment from 'moment-timezone';
 import * as mongoose from 'mongoose';
 
 import authentication from '../middlewares/authentication';
 import permitScopes from '../middlewares/permitScopes';
 import rateLimit from '../middlewares/rateLimit';
 import validator from '../middlewares/validator';
+
+import { search, searchByChevre } from '../service/performance';
+
+const USE_NEW_EVENT_SEARCH = process.env.USE_NEW_EVENT_SEARCH === '1';
 
 const performanceRouter = express.Router();
 
@@ -68,72 +72,74 @@ performanceRouter.get(
             .toDate()
     ],
     validator,
+    // tslint:disable-next-line:cyclomatic-complexity max-func-body-length
     async (req, res, next) => {
         try {
             const countDocuments = req.query.countDocuments === '1';
+            let useLegacySearch = req.query.useLegacySearch === '1';
 
-            // 互換性維持のため
-            // if (typeof req.query.start_from === 'string' && req.query.start_from !== '') {
-            //     req.query.startFrom = moment(req.query.start_from)
-            //         .toDate();
-            // }
-            // if (typeof req.query.start_through === 'string' && req.query.start_through !== '') {
-            //     req.query.startThrough = moment(req.query.start_through)
-            //         .toDate();
-            // }
+            if (!USE_NEW_EVENT_SEARCH) {
+                useLegacySearch = true;
+            }
 
-            // POSへの互換性維持
-            if (req.query.day !== undefined) {
-                if (typeof req.query.day === 'string' && req.query.day.length > 0) {
-                    req.query.startFrom = moment(`${req.query.day}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
-                        .toDate();
-                    req.query.startThrough = moment(`${req.query.day}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
-                        .add(1, 'day')
-                        .toDate();
-
-                    delete req.query.day;
-                }
-
-                if (typeof req.query.day === 'object') {
-                    // day: { '$gte': '20190603', '$lte': '20190802' } } の場合
-                    if (req.query.day.$gte !== undefined) {
-                        req.query.startFrom = moment(`${req.query.day.$gte}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
+            if (useLegacySearch) {
+                // POSへの互換性維持
+                if (req.query.day !== undefined) {
+                    if (typeof req.query.day === 'string' && req.query.day.length > 0) {
+                        req.query.startFrom = moment(`${req.query.day}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
                             .toDate();
-                    }
-                    if (req.query.day.$lte !== undefined) {
-                        req.query.startThrough = moment(`${req.query.day.$lte}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
+                        req.query.startThrough = moment(`${req.query.day}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
                             .add(1, 'day')
                             .toDate();
+
+                        delete req.query.day;
                     }
 
-                    delete req.query.day;
+                    if (typeof req.query.day === 'object') {
+                        // day: { '$gte': '20190603', '$lte': '20190802' } } の場合
+                        if (req.query.day.$gte !== undefined) {
+                            req.query.startFrom = moment(`${req.query.day.$gte}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
+                                .toDate();
+                        }
+                        if (req.query.day.$lte !== undefined) {
+                            req.query.startThrough = moment(`${req.query.day.$lte}T00:00:00+09:00`, 'YYYYMMDDTHH:mm:ssZ')
+                                .add(1, 'day')
+                                .toDate();
+                        }
+
+                        delete req.query.day;
+                    }
                 }
+
+                const conditions: ttts.factory.performance.ISearchConditions = {
+                    ...req.query,
+                    // tslint:disable-next-line:no-magic-numbers
+                    limit: (req.query.limit !== undefined) ? Number(req.query.limit) : 100,
+                    page: (req.query.page !== undefined) ? Math.max(Number(req.query.page), 1) : 1,
+                    sort: (req.query.sort !== undefined) ? req.query.sort : { startDate: 1 },
+                    // POSへの互換性維持のためperformanceIdを補完
+                    ids: (typeof req.query.performanceId === 'string') ? [String(req.query.performanceId)] : undefined
+                };
+
+                const performanceRepo = new ttts.repository.Performance(mongoose.connection);
+
+                let totalCount: number | undefined;
+                if (countDocuments) {
+                    totalCount = await performanceRepo.count(conditions);
+                }
+
+                const performances = await search(conditions)({ performance: performanceRepo });
+
+                if (typeof totalCount === 'number') {
+                    res.set('X-Total-Count', totalCount.toString());
+                }
+
+                res.json({ data: performances });
+            } else {
+                const events = await searchByChevre(req.query)();
+
+                res.json({ data: events });
             }
-
-            const conditions: ttts.factory.performance.ISearchConditions = {
-                ...req.query,
-                // tslint:disable-next-line:no-magic-numbers
-                limit: (req.query.limit !== undefined) ? Number(req.query.limit) : 100,
-                page: (req.query.page !== undefined) ? Math.max(Number(req.query.page), 1) : 1,
-                sort: (req.query.sort !== undefined) ? req.query.sort : { startDate: 1 },
-                // POSへの互換性維持のためperformanceIdを補完
-                ids: (typeof req.query.performanceId === 'string') ? [String(req.query.performanceId)] : undefined
-            };
-
-            const performanceRepo = new ttts.repository.Performance(mongoose.connection);
-
-            let totalCount: number | undefined;
-            if (countDocuments) {
-                totalCount = await performanceRepo.count(conditions);
-            }
-
-            const performances = await ttts.service.performance.search(conditions)({ performance: performanceRepo });
-
-            if (typeof totalCount === 'number') {
-                res.set('X-Total-Count', totalCount.toString());
-            }
-
-            res.json({ data: performances });
         } catch (error) {
             next(error);
         }
@@ -156,7 +162,13 @@ performanceRouter.put(
                         ? { 'ttts_extension.reservationsAtLastUpdateDate': req.body.reservationsAtLastUpdateDate }
                         : undefined,
                     ...(req.body.onlineSalesStatus !== undefined)
-                        ? { 'ttts_extension.online_sales_status': req.body.onlineSalesStatus }
+                        ? {
+                            'ttts_extension.online_sales_status': req.body.onlineSalesStatus,
+                            onlineSalesStatus: req.body.onlineSalesStatus,
+                            eventStatus: (req.body.onlineSalesStatus === ttts.factory.performance.OnlineSalesStatus.Normal)
+                                ? ttts.factory.chevre.eventStatusType.EventScheduled
+                                : ttts.factory.chevre.eventStatusType.EventCancelled
+                        }
                         : undefined,
                     ...(req.body.onlineSalesStatusUpdateUser !== undefined)
                         ? { 'ttts_extension.online_sales_update_user': req.body.onlineSalesStatusUpdateUser }
@@ -168,7 +180,10 @@ performanceRouter.put(
                         }
                         : undefined,
                     ...(req.body.evServiceStatus !== undefined)
-                        ? { 'ttts_extension.ev_service_status': req.body.evServiceStatus }
+                        ? {
+                            'ttts_extension.ev_service_status': req.body.evServiceStatus,
+                            evServiceStatus: req.body.evServiceStatus
+                        }
                         : undefined,
                     ...(req.body.evServiceStatusUpdateUser !== undefined)
                         ? { 'ttts_extension.ev_service_update_user': req.body.evServiceStatusUpdateUser }