+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - パフォーマンス検索を最適化
 - パフォーマンス残席数に、Chevreイベントのオファーごとの集計情報を利用するように調整
+- パフォーマンスからオファー情報を削除
 
 ### Deprecated